+++ test/function/arithmetic/unaryminus.js
@@ -1,4 +1,4 @@
-// test unaryminus
+// test unary minus
 var assert = require('assert');
 var math = require('../../../math.js');
 
@@ -10,21 +10,21 @@ assert.equal(math.eval('4+-2'), 2);
 assert.equal(math.eval('4 + -2'), 2);
 assert.equal(math.eval('4--2'), 6);
 assert.equal(math.eval('4 - -2'), 6);
-assert.equal(math.eval('unaryminus(4)'), -4);
+assert.equal(math.eval('unary(4)'), -4);
 
 // number
-assert.deepEqual(math.unaryminus(2), -2);
-assert.deepEqual(math.unaryminus(-2), 2);
-assert.deepEqual(math.unaryminus(0), 0);
+assert.deepEqual(math.unary(2), -2);
+assert.deepEqual(math.unary(-2), 2);
+assert.deepEqual(math.unary(0), 0);
 
 // complex
-assert.equal(math.unaryminus(math.complex(3, 2)), '-3 - 2i');
-assert.equal(math.unaryminus(math.complex(3, -2)), '-3 + 2i');
-assert.equal(math.unaryminus(math.complex(-3, 2)), '3 - 2i');
-assert.equal(math.unaryminus(math.complex(-3, -2)), '3 + 2i');
+assert.equal(math.unary(math.complex(3, 2)), '-3 - 2i');
+assert.equal(math.unary(math.complex(3, -2)), '-3 + 2i');
+assert.equal(math.unary(math.complex(-3, 2)), '3 - 2i');
+assert.equal(math.unary(math.complex(-3, -2)), '3 + 2i');
 
 // unit
-assert.equal(math.unaryminus(math.unit(5, 'km')).toString(), '-5 km');
+assert.equal(math.unary(math.unit(5, 'km')).toString(), '-5 km');
 
 // string
 assert.throws(function () {math.subtract('hello ', 'world'); });
@@ -33,9 +33,9 @@ assert.throws(function () {math.subtract(123, 'str')});
 
 // array, matrix, range
 a2 = math.matrix([[1,2],[3,4]]);
-var a7 = math.unaryminus(a2);
+var a7 = math.unary(a2);
 assert.ok(a7 instanceof math.type.Matrix);
 assert.deepEqual(a7.size(), [2,2]);
 assert.deepEqual(a7.valueOf(), [[-1,-2],[-3,-4]]);
-assert.deepEqual(math.unaryminus([[1,2],[3,4]]), [[-1,-2],[-3,-4]]);
-assert.deepEqual(math.unaryminus(math.range(1,5)), math.range(-1,-1,-5).valueOf());
+assert.deepEqual(math.unary([[1,2],[3,4]]), [[-1,-2],[-3,-4]]);
+assert.deepEqual(math.unary(math.range(1,5)), math.range(-1,-1,-5).valueOf());